backend/requirements.txt
fastapi[standard]
uvicorn[standard]
asyncpg
redis[hiredis]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
aiogram
python-dotenv
alembic
psycopg2-binary
